Fixes #983, #982, #981, #980
#981: Delete web/lib/formatting.ts (unused number-typed duplicate), update 8 files
to import from @/app/lib/formatting (bigint-typed canonical module).

#982: Delete web/app/lib/dashboard-types.ts, merge all unique types into
web/app/lib/market-types.ts, update 11 importing files.

#980: Delete stacks-api.ts, wallet-service.ts, transaction-service.ts.
Add getMarkets/getTotalVolume to soroban-read-api.ts, migrate all consumers
to soroban-read-api + market-types. Update test mocks.

#983: Consolidate deploy-mainnet.yml and mainnet-deploy.yml into a single
workflow with both workflow_dispatch and tag-push triggers, unified
concurrency group, and all safety checks.
