+++ web/app/components/dashboard/ClaimWinnings.tsx
@@ -3,7 +3,7 @@
 import { useState } from 'react';
 import { Gift, CheckCircle, AlertCircle, RefreshCw, ExternalLink } from 'lucide-react';
 import Link from 'next/link';
-import type { UserBet } from '../../lib/dashboard-types';
+import type { UserBet } from '../../lib/market-types';
 import type { ClaimTxState } from '../../lib/hooks/useClaimWinnings';
 import { formatCurrency } from '../../lib/dashboard-utils';
 import { DEFAULT_NETWORK } from '../../lib/constants';
