+++ web/app/lib/adapters/types.ts
@@ -1,5 +1,5 @@
 /**
  * Domain types for Predinex UI — re-exported from the chain read layer so components
- * do not need to import `stacks-api` directly for types only.
+ * do not need to import deprecated modules directly for types only.
  */
-export type { Pool, ActivityItem } from '../stacks-api';
+export type { Pool, ActivityItem } from '../market-types';
