+++ web/app/lib/disputes/fetchDisputesFromContract.ts
@@ -1,6 +1,6 @@
 import { createScopedLogger } from '@/app/lib/logger';
 const log = createScopedLogger('fetchDisputesFromContract');
-import { getPool } from '../stacks-api';
+import { getPoolFromSoroban } from '../soroban-read-api';
 import type { Dispute } from './types';
 
 export async function fetchDisputesFromContract(): Promise<Dispute[]> {
@@ -17,7 +17,8 @@ export async function fetchDisputesFromContract(): Promise<Dispute[]> {
     for (const event of events) {
       if (event.event === 'smart_contract_event' && event.data.event_name === 'dispute-created') {
         const eventData = event.data.event_data;
-        const pool = await getPool(eventData.pool_id);
+        const poolResult = await getPoolFromSoroban(eventData.pool_id);
+        const pool = poolResult.pool;
 
         disputes.push({
           id: Number(eventData.dispute_id),
