+++ web/app/lib/hooks/useActiveBets.ts
@@ -3,7 +3,7 @@ import { createScopedLogger } from '@/app/lib/logger';
 const log = createScopedLogger('useActiveBets');
 
 import { useState, useCallback } from 'react';
-import { UserBet } from '../dashboard-types';
+import { UserBet } from '../market-types';
 import { getUserBets } from '../dashboard-api';
 import { userDashboardCache } from '../cache-invalidation';
 import { useVisibilityAwarePolling } from './useVisibilityAwarePolling';
