+++ web/tests/components/PoolIntegration.test.tsx
@@ -4,7 +4,7 @@ import { screen, waitFor } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
 import PoolIntegration from '../../app/components/PoolIntegration';
 import * as WalletAdapterProvider from '@/components/WalletAdapterProvider';
-import * as StacksApi from '../../app/lib/stacks-api';
+import * as SorobanReadApi from '../../app/lib/soroban-read-api';
 import * as NetworkMismatch from '../../lib/hooks/useNetworkMismatch';
 import { renderWithProviders } from '../helpers/renderWithProviders';
 
@@ -27,7 +27,7 @@ vi.mock('@/components/WalletAdapterProvider', () => ({
 }));
 
 // Mock stacks-api
-vi.mock('../../app/lib/stacks-api', () => ({
+vi.mock('../../app/lib/soroban-read-api', () => ({
   getMarkets: vi.fn(),
   getPoolCount: vi.fn(),
 }));
@@ -37,7 +37,7 @@ vi.mock('../../lib/hooks/useNetworkMismatch', () => ({
   useNetworkMismatch: vi.fn(),
 }));
 
-const mockPool: StacksApi.Pool = {
+const mockPool: SorobanReadApi.Pool = {
   id: 0,
   title: 'Test Pool',
   description: 'Test Description',
@@ -52,7 +52,7 @@ const mockPool: StacksApi.Pool = {
   status: 'active',
 };
 
-const settledPool: StacksApi.Pool = {
+const settledPool: SorobanReadApi.Pool = {
   ...mockPool,
   id: 1,
   title: 'Settled Pool',
@@ -103,7 +103,7 @@ describe('PoolIntegration', () => {
   });
 
   it('renders loading state initially', () => {
-    vi.mocked(StacksApi.getMarkets).mockImplementation(() => new Promise(() => {})); // Never resolves
+    vi.mocked(SorobanReadApi.getMarkets).mockImplementation(() => new Promise(() => {})); // Never resolves
 
     renderWithProviders(<PoolIntegration />);
 
@@ -111,7 +111,7 @@ describe('PoolIntegration', () => {
   });
 
   it('renders empty state when no pools are available', async () => {
-    vi.mocked(StacksApi.getMarkets).mockResolvedValue([]);
+    vi.mocked(SorobanReadApi.getMarkets).mockResolvedValue([]);
 
     renderWithProviders(<PoolIntegration />);
 
@@ -121,7 +121,7 @@ describe('PoolIntegration', () => {
   });
 
   it('does not contain stale chain references (STX is no longer acceptable)', async () => {
-    vi.mocked(StacksApi.getMarkets).mockResolvedValue([]);
+    vi.mocked(SorobanReadApi.getMarkets).mockResolvedValue([]);
 
     renderWithProviders(<PoolIntegration />);
 
@@ -135,7 +135,7 @@ describe('PoolIntegration', () => {
   });
 
   it('renders pool cards when pools are available', async () => {
-    vi.mocked(StacksApi.getMarkets).mockResolvedValue([mockPool]);
+    vi.mocked(SorobanReadApi.getMarkets).mockResolvedValue([mockPool]);
 
     renderWithProviders(<PoolIntegration />);
 
@@ -151,7 +151,7 @@ describe('PoolIntegration', () => {
   });
 
   it('displays correct pool statistics', async () => {
-    vi.mocked(StacksApi.getMarkets).mockResolvedValue([mockPool, settledPool]);
+    vi.mocked(SorobanReadApi.getMarkets).mockResolvedValue([mockPool, settledPool]);
 
     renderWithProviders(<PoolIntegration />);
 
@@ -174,7 +174,7 @@ describe('PoolIntegration', () => {
   it('still renders the connected wallet action from the current component path', async () => {
     vi.mocked(WalletAdapterProvider.useWallet).mockReturnValue(connectedWallet);
     vi.mocked(NetworkMismatch.useNetworkMismatch).mockReturnValue(mockNetworkMismatch);
-    vi.mocked(StacksApi.getMarkets).mockResolvedValue([mockPool]);
+    vi.mocked(SorobanReadApi.getMarkets).mockResolvedValue([mockPool]);
 
     renderWithProviders(<PoolIntegration />);
 
@@ -192,7 +192,7 @@ describe('PoolIntegration', () => {
   it('enables Place Bet button and navigates to pool detail page when clicked while connected', async () => {
     vi.mocked(WalletAdapterProvider.useWallet).mockReturnValue(connectedWallet);
     vi.mocked(NetworkMismatch.useNetworkMismatch).mockReturnValue(mockNetworkMatch);
-    vi.mocked(StacksApi.getMarkets).mockResolvedValue([mockPool]);
+    vi.mocked(SorobanReadApi.getMarkets).mockResolvedValue([mockPool]);
 
     const user = userEvent.setup();
     renderWithProviders(<PoolIntegration />);
@@ -215,7 +215,7 @@ describe('PoolIntegration', () => {
       connect: mockConnect,
     });
     vi.mocked(NetworkMismatch.useNetworkMismatch).mockReturnValue(mockNetworkMatch);
-    vi.mocked(StacksApi.getMarkets).mockResolvedValue([mockPool]);
+    vi.mocked(SorobanReadApi.getMarkets).mockResolvedValue([mockPool]);
 
     const user = userEvent.setup();
     renderWithProviders(<PoolIntegration />);
@@ -232,7 +232,7 @@ describe('PoolIntegration', () => {
 
   it('shows View Pool Details button when wallet is not connected', async () => {
     vi.mocked(WalletAdapterProvider.useWallet).mockReturnValue(disconnectedWallet);
-    vi.mocked(StacksApi.getMarkets).mockResolvedValue([mockPool]);
+    vi.mocked(SorobanReadApi.getMarkets).mockResolvedValue([mockPool]);
 
     renderWithProviders(<PoolIntegration />);
 
@@ -244,7 +244,7 @@ describe('PoolIntegration', () => {
   });
 
   it('displays settled pool with winner indicator', async () => {
-    vi.mocked(StacksApi.getMarkets).mockResolvedValue([settledPool]);
+    vi.mocked(SorobanReadApi.getMarkets).mockResolvedValue([settledPool]);
 
     renderWithProviders(<PoolIntegration />);
 
@@ -258,7 +258,7 @@ describe('PoolIntegration', () => {
   });
 
   it('calculates and displays correct odds percentages', async () => {
-    vi.mocked(StacksApi.getMarkets).mockResolvedValue([mockPool]);
+    vi.mocked(SorobanReadApi.getMarkets).mockResolvedValue([mockPool]);
 
     renderWithProviders(<PoolIntegration />);
 
@@ -272,7 +272,7 @@ describe('PoolIntegration', () => {
   });
 
   it('handles API errors gracefully', async () => {
-    vi.mocked(StacksApi.getMarkets).mockRejectedValue(new Error('Network error'));
+    vi.mocked(SorobanReadApi.getMarkets).mockRejectedValue(new Error('Network error'));
 
     renderWithProviders(<PoolIntegration />);
 
@@ -284,7 +284,7 @@ describe('PoolIntegration', () => {
   });
 
   it('allows refreshing pools', async () => {
-    vi.mocked(StacksApi.getMarkets).mockResolvedValue([mockPool]);
+    vi.mocked(SorobanReadApi.getMarkets).mockResolvedValue([mockPool]);
 
     const user = userEvent.setup();
     renderWithProviders(<PoolIntegration />);
@@ -296,11 +296,11 @@ describe('PoolIntegration', () => {
     const refreshButton = screen.getByRole('button', { name: /Refresh Pools/i });
     await user.click(refreshButton);
 
-    expect(vi.mocked(StacksApi.getMarkets)).toHaveBeenCalledTimes(2);
+    expect(vi.mocked(SorobanReadApi.getMarkets)).toHaveBeenCalledTimes(2);
   });
 
   it('displays creator address in shortened format', async () => {
-    vi.mocked(StacksApi.getMarkets).mockResolvedValue([mockPool]);
+    vi.mocked(SorobanReadApi.getMarkets).mockResolvedValue([mockPool]);
 
     renderWithProviders(<PoolIntegration />);
 
@@ -314,23 +314,23 @@ describe('PoolIntegration', () => {
   });
 
   it('fetches all pools on mount', async () => {
-    vi.mocked(StacksApi.getMarkets).mockResolvedValue([mockPool]);
+    vi.mocked(SorobanReadApi.getMarkets).mockResolvedValue([mockPool]);
 
     renderWithProviders(<PoolIntegration />);
 
     await waitFor(() => {
-      expect(vi.mocked(StacksApi.getMarkets)).toHaveBeenCalledWith('all');
+      expect(vi.mocked(SorobanReadApi.getMarkets)).toHaveBeenCalledWith('all');
     });
   });
 
   it('handles pools with zero volume correctly', async () => {
-    const emptyPool: StacksApi.Pool = {
+    const emptyPool: SorobanReadApi.Pool = {
       ...mockPool,
       totalA: 0,
       totalB: 0,
     };
 
-    vi.mocked(StacksApi.getMarkets).mockResolvedValue([emptyPool]);
+    vi.mocked(SorobanReadApi.getMarkets).mockResolvedValue([emptyPool]);
 
     renderWithProviders(<PoolIntegration />);
