+++ web/tests/components/mobile-viewport.test.tsx
@@ -14,7 +14,7 @@ import { describe, it, expect, vi, beforeEach, afterEach } from 'vitest';
 import { render, screen, within } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
 import { renderWithProviders } from '../helpers/renderWithProviders';
-import type { Pool } from '../../app/lib/stacks-api';
+import type { Pool } from '../../app/lib/market-types';
 
 // ---------------------------------------------------------------------------
 // Shared mocks
