+++ web/tests/integration/error-states.integration.test.tsx
@@ -94,14 +94,11 @@ vi.mock('../../app/lib/adapters/predinex-contract', () => ({
 }));
 
 vi.mock('../../app/lib/soroban-read-api', () => ({
-  getUserBet: mockGetUserBet,
+  getPoolFromSoroban: mockGetPool,
+  getPoolsBatchFromSoroban: mockGetPools,
   getUserBets: mockGetUserBets,
   getUserActivityFromSoroban: vi.fn().mockResolvedValue([]),
-}));
-
-vi.mock('../../app/lib/stacks-api', () => ({
-  getPools: mockGetPools,
-  getPool: mockGetPool,
+  getPoolCountFromSoroban: vi.fn().mockResolvedValue(0),
 }));
 
 vi.mock('../../hooks/useWallet', () => ({ useWallet: mockUseWallet }));
