+++ web/tests/integration/full-flow.integration.test.tsx
@@ -101,14 +101,12 @@ vi.mock('../../app/lib/adapters/predinex-contract', () => ({
 }));
 
 vi.mock('../../app/lib/soroban-read-api', () => ({
-  getUserBet: mockGetUserBet,
+  getPoolFromSoroban: mockGetPool,
+  getPoolsBatchFromSoroban: mockGetPools,
+  getUserBetFromSoroban: mockGetUserBet,
   getUserBets: mockGetUserBets,
   getUserActivityFromSoroban: vi.fn().mockResolvedValue([]),
-}));
-
-vi.mock('../../app/lib/stacks-api', () => ({
-  getPools: mockGetPools,
-  getPool: mockGetPool,
+  getPoolCountFromSoroban: vi.fn().mockResolvedValue(0),
 }));
 
 vi.mock('../../hooks/useWallet', () => ({ useWallet: mockUseWallet }));
