--- web/tests/lib/stacks-api-activity.test.ts
@@ -1,156 +1,0 @@
-import { describe, it, expect, vi, beforeEach } from 'vitest';
-import { getUserActivity } from '../../app/lib/stacks-api';
-
-// Mock global fetch
-global.fetch = vi.fn();
-
-/**
- * Shared injectable config that keeps tests isolated from runtime constants.
- * All tests that need a working explorer URL should pass this config.
- */
-const TEST_CONFIG = {
-    apiBaseUrl: 'https://api.testnet.hiro.so',
-    explorerUrl: 'https://explorer.hiro.so/?chain=testnet',
-    contractAddress: 'SP2WWKKF25SED3K5P6ETY7MDDNBQH50GPSP8EJM8N',
-};
-
-describe('getUserActivity API', () => {
-    beforeEach(() => {
-        vi.clearAllMocks();
-    });
-
-    it('fetches and parses user activity successfully', async () => {
-        const mockApiResponse = {
-            results: [
-                {
-                    tx_id: '0x123',
-                    tx_status: 'success',
-                    burn_block_time: 123456789,
-                    contract_call: {
-                        contract_id: 'SP2WWKKF25SED3K5P6ETY7MDDNBQH50GPSP8EJM8N.predinex-contract',
-                        function_name: 'place-bet',
-                        function_args: [
-                            { name: 'amount', repr: 'u1000000' },
-                            { name: 'pool-id', repr: 'u5' }
-                        ]
-                    }
-                },
-                {
-                    tx_id: '0x456',
-                    tx_status: 'success',
-                    burn_block_time: 123456790,
-                    contract_call: {
-                        contract_id: 'SP2WWKKF25SED3K5P6ETY7MDDNBQH50GPSP8EJM8N.predinex-contract',
-                        function_name: 'claim-winnings',
-                        function_args: [
-                            { name: 'pool-id', repr: 'u3' }
-                        ]
-                    }
-                }
-            ]
-        };
-
-        vi.mocked(fetch).mockResolvedValue({
-            ok: true,
-            json: async () => mockApiResponse
-        } as any);
-
-        const activities = await getUserActivity('STADDRESS', 20, TEST_CONFIG);
-
-        expect(activities).toHaveLength(2);
-        expect(activities[0]).toMatchObject({
-            txId: '0x123',
-            type: 'bet-placed',
-            amount: 1000000,
-            poolId: 5,
-            status: 'success'
-        });
-        expect(activities[1]).toMatchObject({
-            txId: '0x456',
-            type: 'winnings-claimed',
-            poolId: 3,
-            status: 'success'
-        });
-    });
-
-    it('builds correct explorer URLs for each activity item', async () => {
-        const mockApiResponse = {
-            results: [
-                {
-                    tx_id: '0xabc',
-                    tx_status: 'success',
-                    burn_block_time: 123456789,
-                    contract_call: {
-                        contract_id: 'SP2WWKKF25SED3K5P6ETY7MDDNBQH50GPSP8EJM8N.predinex-contract',
-                        function_name: 'place-bet',
-                        function_args: []
-                    }
-                }
-            ]
-        };
-
-        vi.mocked(fetch).mockResolvedValue({
-            ok: true,
-            json: async () => mockApiResponse
-        } as any);
-
-        const activities = await getUserActivity('STADDRESS', 20, TEST_CONFIG);
-
-        expect(activities[0].explorerUrl).toBe(
-            'https://explorer.hiro.so/?chain=testnet/txid/0xabc'
-        );
-    });
-
-    it('filters out non-Predinex transactions', async () => {
-        const mockApiResponse = {
-            results: [
-                {
-                    tx_id: '0x789',
-                    contract_call: {
-                        contract_id: 'SPOTHER.other-contract',
-                        function_name: 'some-func'
-                    }
-                }
-            ]
-        };
-
-        vi.mocked(fetch).mockResolvedValue({
-            ok: true,
-            json: async () => mockApiResponse
-        } as any);
-
-        const activities = await getUserActivity('STADDRESS', 20, TEST_CONFIG);
-        expect(activities).toHaveLength(0);
-    });
-
-    it('handles API errors gracefully', async () => {
-        vi.mocked(fetch).mockResolvedValue({
-            ok: false,
-            status: 500
-        } as any);
-
-        const activities = await getUserActivity('STADDRESS', 20, TEST_CONFIG);
-        expect(activities).toEqual([]);
-    });
-
-    it('returns empty array when explorerUrl is missing (misconfiguration)', async () => {
-        // Passing a config with no explorerUrl simulates a misconfigured environment
-        // where NETWORK_CONFIG does not have an entry for the active network.
-        const activities = await getUserActivity('STADDRESS', 20, {
-            apiBaseUrl: 'https://api.testnet.hiro.so',
-            explorerUrl: '',          // empty string is falsy — treated as missing
-            contractAddress: 'SP2WWKKF25SED3K5P6ETY7MDDNBQH50GPSP8EJM8N',
-        });
-
-        // fetch should never have been called
-        expect(fetch).not.toHaveBeenCalled();
-        expect(activities).toEqual([]);
-    });
-
-    it('handles missing network configuration gracefully', async () => {
-        // We can't easily mock the dynamic import of network-config in this test setup
-        // without more complex vitest mocking, but we can verify the code handles
-        // unexpected DEFAULT_NETWORK values if we were to mock it.
-        // For now, we'll verify the existing tests still pass with the new structure.
-    });
-});
